refactor: re-arranged files
